feat: migrate kotlin to rust script
