+++ kotlin/README.md
@@ -1,26 +1,35 @@
 # 🧬 Kotlin for Bedrock
 
-This folder contains all relevant support files for Bedrock to work in Kotlin:
+This folder contains all relevant support files for Bedrock to work in Kotlin. **Foreign Tests** are located in `bedrock-tests/src/test/kotlin/bedrock/` folder.
 
-1. Script to cross-compile and build Kotlin bindings.
-2. **Foreign Tests**. Unit tests for Kotlin (`src/test/java/bedrock/` folder). Foreign unit tests run the JUnit test suite via Gradle wrapper.
+All build tasks are exposed as subcommands of `cargo xtask kotlin` (see `xtask/`):
 
-### Building the Kotlin Project
-
-To build the Kotlin project run:
+### Build the Android bindings for release/distribution
 
 ```bash
-    # run from the root project directory
-    ./kotlin/build_kotlin.sh
+cargo xtask kotlin build
 ```
 
-### Running Foreign Tests for Kotlin
+### Run the foreign Kotlin tests
 
 ```bash
-    # run from the root project directory
-    ./kotlin/test_kotlin.sh
+cargo xtask kotlin test
 ```
 
+Builds a host-platform cdylib (so the JVM can load it via JNA) and runs the JUnit suite through the Gradle wrapper.
+
+### Publish a local build to Maven Local
+
+```bash
+cargo xtask kotlin local --version 0.2.10-SNAPSHOT
+```
+
+Builds the Android bindings and runs `publishToMavenLocal`, making the artifact
+available to a consumer project:
+
+```kotlin
+implementation("com.toolsforhumanity:bedrock:0.2.10-SNAPSHOT")
+```
 
 ### Kotlin Project Structure
